Address copilot review comments.

+++ config_src/infra/FMS2/array_mod.F90
@@ -394,7 +394,7 @@ end subroutine dupReal4D
 !< Copy from Fortran array to IntArray_t
 subroutine copy2AInt1D(this,var)
   class(IntArray_t), intent(inout) :: this  !< The destination array container
-  real, dimension(:), intent(in) :: var      !< The source Fortran array
+  integer, dimension(:), intent(in) :: var      !< The source Fortran array
 
   this%data(:) = var(:)
 
@@ -403,7 +403,7 @@ end subroutine copy2AInt1D
 !< Copy from Fortran array to IntArray_t
 subroutine copy2AInt2D(this,var)
   class(IntArray_t), intent(inout) :: this  !< The destination array container
-  real, dimension(:,:), intent(in) :: var    !< The source Fortran array
+  integer, dimension(:,:), intent(in) :: var    !< The source Fortran array
 
   ! Local variables
   integer :: i, j, n1, n2,idx
@@ -422,7 +422,7 @@ end subroutine copy2AInt2D
 !< Copy from Fortran array to IntArray_t
 subroutine copy2AInt3D(this,var)
   class(IntArray_t), intent(inout) :: this  !< The destination array container
-  real, dimension(:,:,:), intent(in) :: var  !< The source Fortran array
+  integer, dimension(:,:,:), intent(in) :: var  !< The source Fortran array
 
   ! Local variables
   integer :: i, j, k, n1, n2, n3, idx
@@ -444,7 +444,7 @@ end subroutine copy2AInt3D
 !< Copy from Fortran array to IntArray_t
 subroutine copy2AInt4D(this,var)
   class(IntArray_t), intent(inout) :: this   !< The destination array container
-  real, dimension(:,:,:,:), intent(in) :: var !< The source Fortran array
+  integer, dimension(:,:,:,:), intent(in) :: var !< The source Fortran array
 
   ! Local variables
   integer :: i, j, k, m, n1, n2, n3, n4, idx
@@ -469,7 +469,7 @@ end subroutine copy2AInt4D
 ! Copy from 1D IntArray_t to Fortran
 subroutine copy2FInt1D(this,var)
   class(IntArray_t), intent(in) :: this    !< The source array container
-  real, dimension(:), intent(inout) :: var  !< The destination Fortran array
+  integer, dimension(:), intent(inout) :: var  !< The destination Fortran array
 
   var(:) = this%data(:)
 
@@ -478,7 +478,7 @@ end subroutine copy2FInt1D
 ! Copy from 2D IntArray_t to Fortran
 subroutine copy2FInt2D(this,var)
   class(IntArray_t), intent(in) :: this     !< The source array container
-  real, dimension(:,:), intent(inout) :: var !< The destination Fortran array
+  integer, dimension(:,:), intent(inout) :: var !< The destination Fortran array
 
   ! Local variables
   integer :: i, j, n1,n2,idx
@@ -497,7 +497,7 @@ end subroutine copy2FInt2D
 ! Copy from 3D IntArray_t to Fortran
 subroutine copy2FInt3D(this,var)
   class(IntArray_t), intent(in) :: this       !< The source array container
-  real, dimension(:,:,:), intent(inout) :: var !< The destination Fortran array
+  integer, dimension(:,:,:), intent(inout) :: var !< The destination Fortran array
 
   ! Local variables
   integer :: i, j, k, n1,n2,n3, idx
@@ -519,7 +519,7 @@ end subroutine copy2FInt3D
 ! Copy from 4D IntArray_t to Fortran
 subroutine copy2FInt4D(this,var)
   class(IntArray_t), intent(in) :: this          !< The source array container
-  real, dimension(:,:,:,:), intent(inout) :: var  !< The destination Fortran array
+  integer, dimension(:,:,:,:), intent(inout) :: var  !< The destination Fortran array
 
   ! Local variables
   integer :: i, j, k, m, n1, n2, n3, n4, idx
@@ -543,7 +543,7 @@ end subroutine copy2FInt4D
 
 subroutine dupInt1D(this,var)
   class(IntArray_t), intent(inout) :: this !< The resulting array container
-  real, dimension(:) :: var                 !< The Fortran array to duplicate
+  integer, dimension(:) :: var                 !< The Fortran array to duplicate
 
   ! Local variables
   integer, dimension(1) :: lb, ub
@@ -556,7 +556,7 @@ end subroutine dupInt1D
 
 subroutine dupInt2D(this,var)
   class(IntArray_t), intent(inout) :: this !< The resulting array container
-  real, dimension(:,:) :: var               !< The Fortran array to duplicate
+  integer, dimension(:,:) :: var               !< The Fortran array to duplicate
 
   ! Local variables
   integer, dimension(2) :: lb, ub
@@ -571,7 +571,7 @@ end subroutine dupInt2D
 
 subroutine dupInt3D(this,var)
   class(IntArray_t), intent(inout) :: this !< The resulting array container
-  real, dimension(:,:,:) :: var             !< The Fortran array to duplicate
+  integer, dimension(:,:,:) :: var             !< The Fortran array to duplicate
 
   ! Local variables
   integer, dimension(3) :: lb, ub
@@ -588,7 +588,7 @@ end subroutine dupInt3D
 
 subroutine dupInt4D(this,var)
   class(IntArray_t), intent(inout) :: this !< The resulting array container
-  real, dimension(:,:,:,:) :: var           !< The Fortran array to duplicate
+  integer, dimension(:,:,:,:) :: var           !< The Fortran array to duplicate
 
   ! Local variables
   integer, dimension(4) :: lb, ub
